Refactor commercialRegistrations route structure
Reorganize commercialRegistrations by moving type definitions to a separate file and updating imports accordingly. This improves code maintainability and readability by separating route logic from type definitions.
