+++ src/routes/commercialRegistrations/commercialRegistrations.ts
@@ -1,71 +1,12 @@
 import type { Request, Response } from "express";
-import { businesses } from "../fixtures/businesses";
-
-export type SearchByNameResponseBody = {
-  name: string;
-  registration_number: string;
-  /**
-   * Only required for companies in Germany
-   */
-  registration_issuer?: string;
-};
-
-export type Country = string | "DE";
-export type LegalForm = string | "GMBH";
-
-export type BusinessAddress = {
-  country: Country;
-  postal_code: string;
-  city: string;
-  line_1: string;
-  line_2: string;
-};
-
-export type SearchByNameQuery = {
-  country: string;
-  name: string;
-};
-
-export type LegalRepresentative = {
-  first_name: string;
-  last_name: string;
-};
-
-/**
- * This is used for samples
- */
-export type Business = {
-  name: string;
-  address: BusinessAddress;
-  legal_form: LegalForm;
-  tax_country: Country;
-  registration_number: string;
-  registration_issuer: string;
-  /**
-   * YYYY-MM_DD
-   */
-  registration_date: string;
-  /**
-   * YYYY-MM_DD
-   */
-  registry_updated_at: string;
-  legal_representatives: LegalRepresentative[];
-  commercial_registry_industry_key: string[];
-};
-
-export type findQuery = {
-  registration_number: string;
-  registration_issuer: string;
-  country: Country;
-};
-
-export type ModelNotFoundError = {
-  title: string | "Model Not Found";
-  status: string | "404";
-  id: string;
-  detail: string;
-  code: string | "model_not_found";
-};
+import { businesses } from "../../fixtures/businesses";
+import {
+  SearchQuery,
+  SearchByNameResponseBody,
+  findQuery,
+  ModelNotFoundError,
+  Business,
+} from "./types";
 
 /**
  * @see {@link https://docs.solarisgroup.com/api-reference/onboarding/businesses/#tag/Business-Registrations/paths/~1v1~1commercial_registrations~1search_by_name/get}
@@ -75,7 +16,7 @@ export type ModelNotFoundError = {
  * Returns a mock business if more than 2 characters are provided for country and name
  */
 export const searchByName = async (
-  req: Request<{}, {}, {}, SearchByNameQuery>,
+  req: Request<{}, {}, {}, SearchQuery>,
   res: Response<SearchByNameResponseBody | ModelNotFoundError>
 ) => {
   const { country = "DE", name } = req.query;
